backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
opencv-python-headless==4.9.0.80
pytesseract==0.3.10
pdfplumber==0.10.3
gunicorn==21.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
